+++ cmd/pint/scan.go
@@ -70,7 +70,8 @@ func checkRules(ctx context.Context, workers int, isOffline bool, gen *config.Pr
 			default:
 				if entry.Rule.RecordingRule != nil {
 					rulesParsedTotal.WithLabelValues(config.RecordingRuleType).Inc()
-					slog.LogAttrs(ctx, slog.LevelDebug, "Found recording rule",
+					slog.LogAttrs(
+						ctx, slog.LevelDebug, "Found recording rule",
 						slog.String("path", entry.Path.Name),
 						slog.String("record", entry.Rule.RecordingRule.Record.Value),
 						slog.String("lines", entry.Rule.Lines.String()),
@@ -79,7 +80,8 @@ func checkRules(ctx context.Context, workers int, isOffline bool, gen *config.Pr
 				}
 				if entry.Rule.AlertingRule != nil {
 					rulesParsedTotal.WithLabelValues(config.AlertingRuleType).Inc()
-					slog.LogAttrs(ctx, slog.LevelDebug, "Found alerting rule",
+					slog.LogAttrs(
+						ctx, slog.LevelDebug, "Found alerting rule",
 						slog.String("path", entry.Path.Name),
 						slog.String("alert", entry.Rule.AlertingRule.Alert.Value),
 						slog.String("lines", entry.Rule.Lines.String()),
@@ -87,7 +89,8 @@ func checkRules(ctx context.Context, workers int, isOffline bool, gen *config.Pr
 					)
 				}
 				if entry.Rule.Error.Err != nil {
-					slog.LogAttrs(ctx, slog.LevelDebug, "Found invalid rule",
+					slog.LogAttrs(
+						ctx, slog.LevelDebug, "Found invalid rule",
 						slog.String("path", entry.Path.Name),
 						slog.String("lines", entry.Rule.Lines.String()),
 						slog.String("state", entry.State.String()),
@@ -127,7 +130,8 @@ func checkRules(ctx context.Context, workers int, isOffline bool, gen *config.Pr
 	}
 	for _, pd := range summary.GetPrometheusDetails() {
 		for _, dc := range pd.DisabledChecks {
-			slog.LogAttrs(ctx, slog.LevelWarn,
+			slog.LogAttrs(
+				ctx, slog.LevelWarn,
 				"Some checks were disabled because configured server doesn't seem to support all Prometheus APIs",
 				slog.String("prometheus", pd.Name),
 				slog.String("api", dc.API),
@@ -154,7 +158,8 @@ func scanWorker(ctx context.Context, jobs <-chan scanJob, results chan<- reporte
 			return
 		default:
 			if job.entry.State == discovery.Unknown {
-				slog.LogAttrs(ctx, slog.LevelWarn,
+				slog.LogAttrs(
+					ctx, slog.LevelWarn,
 					"Bug: unknown rule state",
 					slog.String("path", job.entry.Path.String()),
 					slog.Int("line", job.entry.Rule.Lines.First),