+++ internal/checks/promql_vector_matching.go
@@ -179,7 +179,8 @@ func (c VectorMatchingCheck) checkNode(ctx context.Context, rule parser.Rule, ex
 							{
 								Message: fmt.Sprintf(
 									"Using `on(%s)` won't produce any results on %s because [series returned by this query](%s) doesn't have the `%s` label.",
-									name, promText(c.prom.Name(), qr.URI), link, name),
+									name, promText(c.prom.Name(), qr.URI), link, name,
+								),
 								Pos:         expr.Value.Pos,
 								FirstColumn: int(pos.Start) + 1,
 								LastColumn:  int(pos.End),