+++ internal/reporter/github_test.go
@@ -951,10 +951,49 @@ func TestGitHubReporter(t *testing.T) {
 			timeout:     time.Second,
 			httpHandler: http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 				if r.Method == http.MethodPost && r.URL.Path == "/api/v3/repos/foo/bar/pulls/123/reviews" {
+					expected := `### This pull request was validated by [pint](https://github.com/cloudflare/pint).
+:heavy_exclamation_mark:	Problems found.
+| Severity | Number of problems |
+| --- | --- |
+| Fatal | 1 |
+<details><summary>Stats</summary>
+<p>
+
+| Stat | Value |
+| --- | --- |
+| Version | v0.0.0 |
+| Number of rules parsed | 0 |
+| Number of rules checked | 0 |
+| Number of problems found | 1 |
+| Number of offline checks | 0 |
+| Number of online checks | 0 |
+| Checks duration | 0 |
+
+</p>
+</details>
+
+<details><summary>Problems</summary>
+<p>
+
+` + "```" + `
+Fatal: syntax error (mock)
+  ---> foo.txt:1 -> ` + "`target is down`" + `
+       ^^^ syntax details
+
+` + "```" + `
+
+</p>
+</details>
+
+`
 					body, _ := io.ReadAll(r.Body)
-					b := strings.TrimSpace(strings.TrimRight(string(body), "\n\t\r"))
-					if b != `{"commit_id":"HEAD","body":"### This pull request was validated by [pint](https://github.com/cloudflare/pint).\n:heavy_exclamation_mark:\tProblems found.\n| Severity | Number of problems |\n| --- | --- |\n| Fatal | 1 |\n<details><summary>Stats</summary>\n<p>\n\n| Stat | Value |\n| --- | --- |\n| Version | v0.0.0 |\n| Number of rules parsed | 0 |\n| Number of rules checked | 0 |\n| Number of problems found | 1 |\n| Number of offline checks | 0 |\n| Number of online checks | 0 |\n| Checks duration | 0 |\n\n</p>\n</details>\n\n<details><summary>Problems</summary>\n<p>\n\nFailed to generate list of problems: open foo.txt: no such file or directory\n</p>\n</details>\n\n","event":"COMMENT"}` {
-						t.Errorf("Unexpected comment: %s", b)
+					type jr struct {
+						Body string
+					}
+					var j jr
+					_ = json.Unmarshal(body, &j)
+					if diff := cmp.Diff(expected, j.Body); diff != "" {
+						t.Errorf("Unexpected review body (-want +got):\n%s", diff)
 						t.FailNow()
 					}
 				}